requirements.txt
streamlit>=1.33
pandas>=2.0
requests>=2.31
supabase>=2.0
coinbase-advanced-py>=1.8
